wip: split vault status adapter
